+++ lib/components/button/toggle.dart
@@ -1,5 +1,6 @@
 import 'dart:html';
 import "package:angular/angular.dart";
+import 'package:angular_forms/angular_forms.dart';
 
 @Directive(
     selector: "bs-toggle-button",