+++ lib/components/datepicker/index.dart
@@ -6,6 +6,7 @@ import 'package:intl/intl.dart';
 import 'package:node_shims/js.dart';
 import 'package:ng_bootstrap/components/dropdown/index.dart';
 import 'package:ng_bootstrap/components/button/toggle.dart';
+import 'package:angular_forms/angular_forms.dart';
 
 part "date_picker.dart";
 part "date_picker_popup.dart";