+++ lib/components/validators/min_length_validator.dart
@@ -1,5 +1,6 @@
 import 'package:angular/angular.dart';
 import 'package:angular/core.dart';
+import 'package:angular_forms/angular_forms.dart';
 
 /// A [Directive] adding minimum-length validator to controls with `minlength`.
 ///