+++ lib/ng_bootstrap.dart
@@ -1,26 +1,26 @@
-import 'package:ng_bootstrap/components/accordion/accordion.dart';
-import 'package:ng_bootstrap/components/alert/alert.dart';
-import 'package:ng_bootstrap/components/button/radio.dart';
-import 'package:ng_bootstrap/components/button/toggle.dart';
-import 'package:ng_bootstrap/components/carousel/carousel.dart';
-import 'package:ng_bootstrap/components/collapse/collapse.dart';
-import 'package:ng_bootstrap/components/datepicker/index.dart';
-import 'package:ng_bootstrap/components/dropdown/index.dart';
-import 'package:ng_bootstrap/components/file_upload/file_upload.dart';
-import 'package:ng_bootstrap/components/input/input.dart';
-import 'package:ng_bootstrap/components/modal/modal.dart';
-import 'package:ng_bootstrap/components/pagination/pager.dart';
-import 'package:ng_bootstrap/components/pagination/pagination.dart';
-import 'package:ng_bootstrap/components/progress/progress.dart';
-import 'package:ng_bootstrap/components/table/table_directives.dart';
-import 'package:ng_bootstrap/components/tabs/tabs.dart';
-import 'package:ng_bootstrap/components/tabsx/tabsx.dart';
-import 'package:ng_bootstrap/components/tooltip/tooltip.dart';
-import 'package:ng_bootstrap/components/typeahead/typeahead.dart';
+import 'components/accordion/accordion.dart';
+import 'components/alert/alert.dart';
+import 'components/button/radio.dart';
+import 'components/button/toggle.dart';
+import 'components/carousel/carousel.dart';
+import 'components/collapse/collapse.dart';
+import 'components/datepicker/index.dart';
+import 'components/dropdown/index.dart';
+import 'components/file_upload/file_upload.dart';
+import 'components/input/input.dart';
+import 'components/modal/modal.dart';
+import 'components/pagination/pager.dart';
+import 'components/pagination/pagination.dart';
+import 'components/progress/progress.dart';
+import 'components/table/table_directives.dart';
+import 'components/tabs/tabs.dart';
+import 'components/tabsx/tabsx.dart';
+import 'components/tooltip/tooltip.dart';
+import 'components/typeahead/typeahead.dart';
 import 'components/validators/min_length_validator.dart';
 import 'components/validators/max_length_validator.dart';
 
-export 'package:ng_bootstrap/core/position.dart';
+export 'core/position.dart';
 
 export 'components/accordion/accordion.dart';
 export 'components/alert/alert.dart';
@@ -43,6 +43,7 @@ export 'components/tooltip/tooltip.dart';
 export 'components/typeahead/typeahead.dart';
 export 'components/validators/min_length_validator.dart';
 export 'components/validators/max_length_validator.dart';
+export 'components/prompt/prompt_service.dart';
 
 const List BS_DIRECTIVES = const [
   BsAccordionComponent,