+++ tool/build.dart
@@ -4,5 +4,5 @@ import 'package:build_runner/build_runner.dart';
 import 'package:sass_builder/phase.dart';
 
 Future main() async {
-  await build(new PhaseGroup()..addPhase(sassPhase), deleteFilesByDefault: true);
+  await build([sassBuildAction], deleteFilesByDefault: true);
 }